backend/inpainting/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
